requirements.txt
dbus-python >= 1.4.0, < 2.0.0
PyGObject >= 3.52.0, < 4.0.0
